assert: move Test_indentMessageLines into a separate file
